# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Pesantren Tahfidz Qur'an dan Digital Arrahman API",
    "description": "Backend API for Pesantren Tahfidz Qur'an dan Digital Ar-Rahman",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "paths": {
...
    "/users": {
      "get": {
        "summary": "List Users",
        "tags": ["Users"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "User list" }
        }
      },
      "post": {
        "summary": "Create User",
        "tags": ["Users"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "201": { "description": "User created" }
        }
      }
    },
...
    "/roles": {
      "get": {
        "summary": "List Roles",
        "tags": ["Roles"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Role list" }
        }
      },
      "post": {
        "summary": "Create Role",
        "tags": ["Roles"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "201": { "description": "Role created" }
        }
      }
    },
...
    "/spp": {
      "get": {
        "summary": "List SPP",
        "tags": ["SPP"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "SPP list" }
        }
      }
    },
    "/students": {
      "get": {
        "summary": "List Students",
        "tags": ["Students"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Student list" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}